requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.5
langchain-core>=0.1.16,<0.2
langchain-community==0.0.10
docx2txt==0.8
pypdf==3.17.1
chromadb==0.4.22
sentence-transformers==2.2.2
openai>=1.10.0,<2.0.0
python-dotenv==1.0.0
streamlit==1.29.0